youtube_video_to_blog/requirements.txt
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain-huggingface
langchain_groq
langgraph
langgraph-cli[inmem]
youtube_transcript_api

